requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.121.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
lxml==6.0.2
openai==2.7.1
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.6
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
